Add files via upload
优化脚本部分内容

+++ auto.sh
@@ -1,12 +1,8 @@
 #!/bin/bash
-cd /root
 
-# 使用 curl 下载文件，并跟随重定向
-curl -L -o Dialer.zip https://github.com/dogliu666/ESurfingDialer-For-Docker/releases/download/Latest/Dialer.zip
-
-# 检查文件是否下载成功
-if [ $? -ne 0 ]; then
-    echo "下载 Dialer.zip 文件失败"
+# 检查文件是否存在
+if [ ! -f "Dialer.zip" ]; then
+    echo "Dialer.zip 文件不存在"
     exit 1
 fi
 
@@ -23,22 +19,44 @@ if ! unzip -tq Dialer.zip; then
 fi
 
 # 解压缩 dialer.zip 文件
-unzip /root/Dialer.zip
-cd /root/Dialer
+unzip -o Dialer.zip -d /root/Dialer
+if [ $? -ne 0 ]; then
+    echo "解压缩 Dialer.zip 失败"
+    exit 1
+fi
+cd /root/Dialer || { echo "目录 /root/Dialer 不存在"; exit 1; }
 
 # 构建 Docker 容器
 docker build -t Dialer .
+if [ $? -ne 0 ]; then
+    echo "Docker 容器构建失败"
+    exit 1
+fi
 
 # 导出 Docker 镜像
 docker save -o Dialer.tar Dialer
+if [ $? -ne 0 ]; then
+    echo "Docker 镜像导出失败"
+    exit 1
+fi
 
-# 加载镜像
+# 加载 Docker 镜像
 docker load -i ./Dialer.tar
+if [ $? -ne 0 ]; then
+    echo "Docker 镜像加载失败"
+    exit 1
+fi
 
 # 输入 校园网 账密
 read -p "请输入账号: " account
-read -p "请输入密码: " pwd
+read -sp "请输入密码: " pwd
 echo
+
+# 验证输入的账号和密码
+if [ -z "$account" ] || [ -z "$pwd" ]; then
+    echo "账号或密码不能为空"
+    exit 1
+fi
 
 # 将输入的 账密 保存到 Config.txt 中
 echo "account=$account" > Config.txt
@@ -48,4 +66,8 @@ echo "pwd=$pwd" >> Config.txt
 source ./Config.txt
 
 # 运行 docker run 命令
-docker run -itd -e DIALER_USER="$account" -e DIALER_PASSWORD="$pwd" --name dialer-client --network host --restart=always dialer
+docker run -itd -e DIALER_USER="$account" -e DIALER_PASSWORD="$pwd" --name dialer-client --network host --restart=always Dialer
+if [ $? -ne 0 ]; then
+    echo "Docker 容器运行失败"
+    exit 1
+fi
